[hist] Cast benchmark ranges
This avoids sign conversion warnings.

+++ hist/histv7/benchmark/hist_benchmark_stats.cxx
@@ -3,6 +3,7 @@
 
 #include <benchmark/benchmark.h>
 
+#include <cstddef>
 #include <random>
 #include <vector>
 
@@ -18,7 +19,7 @@ struct RHistStats1 : public benchmark::Fixture {
    {
       std::mt19937 gen;
       std::uniform_real_distribution<> dis;
-      fNumbers.resize(state.range(0));
+      fNumbers.resize(static_cast<std::size_t>(state.range(0)));
       for (std::size_t i = 0; i < fNumbers.size(); i++) {
          fNumbers[i] = dis(gen);
       }
@@ -57,7 +58,7 @@ struct RHistStats2 : public benchmark::Fixture {
    {
       std::mt19937 gen;
       std::uniform_real_distribution<> dis;
-      fNumbers.resize(2 * state.range(0));
+      fNumbers.resize(2 * static_cast<std::size_t>(state.range(0)));
       for (std::size_t i = 0; i < fNumbers.size(); i++) {
          fNumbers[i] = dis(gen);
       }
